refactor validation
fixes idrinth-api-bench/issues#565
