--- src/locales/en.ts
@@ -1,50 +1,0 @@
-/* eslint max-len:0 */
-const lang = {
-  no_tasks: 'Can\'t measure zero tasks.',
-  duplicate_task_id: 'The id %1% is shared.',
-  invalid_key: 'The given key %1% is invalid.',
-  finished_analyzing: 'Analyzation of %1% finished.',
-  initialization: 'Initializing tasks to have %1% Threads x %2% Repetitions todos',
-  starting_result: 'Starting supplied result handler.',
-  data: 'Data',
-  done: 'Done',
-  starting_workers: 'Starting up %1% Workers.',
-  starting_after: 'Starting after request.',
-  no_after: 'No after request, done.',
-  finished_requests: 'Finished requesting all %1%.',
-  starting_analyzation: 'Starting analysation of %1%.',
-  next_request: 'Starting next request.',
-  starting_validation: 'Starting validation of %1%.',
-  end_thread: 'All requests done, terminating thread.',
-  after_done: 'After done.',
-  no_response_status: 'Request returned no status',
-  response_status_above_2xx: 'Request returned status above 200-299 range',
-  response_status_below_2xx: 'Request returned status below 200-299 range',
-  response_status_not_404: 'Request returned status %1%, not 404',
-  response_status_not_403: 'Request returned status %1%, not 403',
-  variable_default_value: 'Can\'t handle variable default value on %1%.',
-  name_invalid: 'Invalid project name.',
-  already_created: 'Project folder already exists.',
-  no_content_type: 'The content-type header is missing.',
-  no_json_content_type: 'The content-type %1% is not application/json.',
-  invalid_json_body: 'The JSON body is invalid. %1%',
-  openapi_invalid: 'Your Open-Api document is not valid, please check for errors!',
-  no_openapi_document_given: 'You need to provide an Open-Api document path.',
-  no_xml_content_type: 'The content-type application/jason is not */xml.',
-  invalid_xml_body: 'The XML body is invalid.',
-  too_slow: 'The response time was above %1% ns.',
-  maximum_below_threads: 'Maximum is lower than the initial thread count.',
-  increment_below_one: 'You need to increment your threads each iteration, currently increment is below one.',
-  maximum_below_one: 'You can not set the maximum below one.',
-  server_header_is_set: 'The header Server is set. Remove this for improved security.',
-  powered_by_is_set: 'The header X-Powered-By is set. It shares critical information with the world.',
-  validation_errors: 'There were %1% errors encountered.',
-  validation_warnings: 'There were %1% warnings encountered.',
-  invalid_pre_definition: 'The definition of pre on route %1% is not valid.',
-  invalid_post_definition: 'The definition of post on route %1% is not valid.',
-  unknown_route_property: 'The property %1% on route %1% is unknown.',
-  no_har_document_given: 'The  path didn\'t point to an HAR document.',
-  cant_determine_valid_type: 'Failed to find a supported file type.',
-};
-
-export default lang;
