--- src/locales/es.ts
@@ -1,38 +1,0 @@
-/* eslint max-len:0 */
-const lang = {
-  no_tasks: 'No es posible medir ninguna tarea.',
-  duplicate_task_id: 'El id %1% está duplicado.',
-  invalid_key: 'La llave %1% es inválida.',
-  finished_analyzing: 'Análisis de %1% finalizado.',
-  initialization: 'Inicializando tareas para tener %1% Hilos x %2% repeticiones',
-  starting_result: 'Iniciando el controlador de resultados proporcionados.',
-  data: 'Datos',
-  done: 'Listos',
-  starting_workers: 'Empezando %1% Workers.',
-  starting_after: 'Comenzando después de la solicitud.',
-  no_after: 'Listo, no hay solicitudes posteriores.',
-  finished_requests: 'Finalizado de solicitar %1%.',
-  starting_analyzation: 'Comenzando análisis de %1%.',
-  next_request: 'Comenzando la siguiente solicitud.',
-  starting_validation: 'Iniciando validacion de %1%.',
-  end_thread: 'Todas las solicitudes estan listas, finalizando hilo.',
-  after_done: 'Después de finalizado.',
-  no_response_status: 'La respuesta no contiene estado',
-  response_status_above_2xx: 'La respuesta retornó un estado arriba del rango 200-299',
-  response_status_below_2xx: 'La respuesta retornó un estado abajo del rango 200-299',
-  response_status_not_404: 'El estado de la respuesta es %1%, no 404',
-  response_status_not_403: 'El estado de la respuesta es %1%, no 403',
-  variable_default_value: 'No se puede manejar el valor por defecto %1%.',
-  name_invalid: 'Nombre de proyecto inválido.',
-  already_created: 'Directo de proyecto ya existe.',
-  no_content_type: 'Falta el header content-type.',
-  no_json_content_type: 'El content-type %1% no es application/json.',
-  invalid_json_body: 'El cuerpo del JSON body es inválido. %1%',
-  openapi_invalid: 'Su documento Open-Api no es válido, verifique si hay errores!',
-  no_openapi_document_given: 'Debes proporcionar una ruta de documento Open-Api.',
-  no_xml_content_type: 'El content-type %1% no es */xml.',
-  invalid_xml_body: 'El cuerpo XML es inválido.',
-  too_slow: 'El tiempo de respuesta fue superior a %1% ns.',
-};
-
-export default lang;
