--- src/locales/language-key.ts
@@ -1,3 +1,0 @@
-/* eslint max-len:0 */
-type languageKey = 'no_tasks'|'duplicate_task_id'|'invalid_key'|'finished_analyzing'|'initialization'|'starting_result'|'data'|'done'|'starting_workers'|'starting_after'|'no_after'|'finished_requests'|'starting_analyzation'|'next_request'|'starting_validation'|'end_thread'|'after_done'|'no_response_status'|'response_status_above_2xx'|'response_status_below_2xx'|'response_status_not_404'|'response_status_not_403'|'variable_default_value'|'name_invalid'|'already_created'|'no_content_type'|'no_json_content_type'|'invalid_json_body'|'openapi_invalid'|'no_openapi_document_given'|'no_xml_content_type'|'invalid_xml_body'|'too_slow'|'maximum_below_threads'|'increment_below_one'|'maximum_below_one'|'server_header_is_set'|'powered_by_is_set'|'validation_errors'|'validation_warnings'|'invalid_pre_definition'|'invalid_post_definition'|'unknown_route_property'|'no_har_document_given'|'cant_determine_valid_type';
-export default languageKey;
