--- src/locales/languages.ts
@@ -1,7 +1,0 @@
-/* eslint max-len:0 */
-const languages = [
-  'de',
-  'en',
-  'es',
-];
-export default languages;
