+++ src/validate-tasks.ts
@@ -1,18 +1,10 @@
-import Task from './task';
-import language from './helper/language';
+import Task from './task.js';
+import language from './helper/language.js';
 import {
   EMPTY,
-} from './constants';
+} from './constants.js';
+import noDuplicateIds from './no-duplicate-ids.js';
 
-const noDuplicateIDs = (tasks: Array<Task>,) => {
-  const ids: Array<string> = [];
-  for (const task of tasks) {
-    if (ids.includes(task.id,)) {
-      throw new Error(language('duplicate_task_id', task.id,),);
-    }
-    ids.push(task.id,);
-  }
-};
 const executableAmount = (
   repetitions: number,
   threads: number,
@@ -29,5 +21,5 @@ export default function validateTasks(
   tasks: Array<Task>,
 ): void {
   executableAmount(repetitions, threads, tasks,);
-  noDuplicateIDs(tasks,);
+  noDuplicateIds(tasks,);
 }
